Module17-FileUploadsDownloads is complete

+++ Module17-FileUploadsDownloads/controllers/admin.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator");
 const Product = require("../models/product");
+const fileHelper = require("../util/file");
 
 module.exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -15,7 +16,26 @@ module.exports.getAddProduct = (req, res, next) => {
 
 module.exports.postAddProduct = async (req, res, next) => {
   console.log("postAddProduct");
-  const { title, imageUrl, price, description } = req.body;
+  const { title, price, description } = req.body;
+  const image = req.file;
+  console.log(image);
+
+  if (!image) {
+    return res.status(422).render("admin/edit-product", {
+      path: "/admin/edit-product",
+      pageTitle: "Add Product",
+      editing: false,
+      hasError: true,
+      product: {
+        title,
+        price,
+        description,
+      },
+      isAuthenticated: req.session.isLoggedIn,
+      errorMessage: "Attached file is not an image.",
+      validationErrors: [],
+    });
+  }
 
   const errors = validationResult(req);
   console.log("errors", errors);
@@ -28,7 +48,6 @@ module.exports.postAddProduct = async (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description,
       },
@@ -43,7 +62,7 @@ module.exports.postAddProduct = async (req, res, next) => {
       title,
       price,
       description,
-      imageUrl,
+      imageUrl: image.path,
       // FIXME: Focus there, Your can pass req.session.user directly, mongoose can deal with this, just add user._id
       userId: req.session.user,
     });
@@ -51,6 +70,7 @@ module.exports.postAddProduct = async (req, res, next) => {
     res.redirect("/");
   } catch (err) {
     const error = new Error(err);
+    console.log(err);
     error.httpStatusCode = 500;
     return next(error);
   }
@@ -85,7 +105,8 @@ module.exports.getEditProduct = async (req, res, next) => {
 };
 
 module.exports.postEditProduct = async (req, res, next) => {
-  const { prodId, title, price, imageUrl, description } = req.body;
+  const { prodId, title, price, description } = req.body;
+  const image = req.file;
 
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
@@ -96,7 +117,6 @@ module.exports.postEditProduct = async (req, res, next) => {
       hasError: true,
       product: {
         title,
-        imageUrl,
         price,
         description,
         prodId,
@@ -116,7 +136,10 @@ module.exports.postEditProduct = async (req, res, next) => {
 
     product.title = title;
     product.price = price;
-    product.imageUrl = imageUrl;
+    if (image) {
+      fileHelper.deleteFile(product.imageUrl);
+      product.imageUrl = image.path;
+    }
     product.description = description;
     await product.save();
     res.redirect("/admin/products");
@@ -131,6 +154,12 @@ module.exports.postDeleteProduct = async (req, res, next) => {
   const { prodId } = req.body;
 
   try {
+    const product = await Product.findById(prodId);
+    if (!product) {
+      return next(new Error("Product not found."));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+
     // FIXME: current version mongoose already change the `findByIdAndRemove` to `findByIdAndDelete`, findByIdAndRemove has deprecated;
     await Product.deleteOne({
       _id: prodId,
